Be consistent with what we're doing elsewhere and rename the `cmd` package to `cli`
